Add new helper function for handling Omni Stop Loss data
