+++ features/omni-kit/automation/helpers/automationFormatters.ts
@@ -1,5 +1,6 @@
 import type BigNumber from 'bignumber.js'
-import { formatAmount, formatPercent } from 'helpers/formatters/format'
+import { lambdaPercentageDenomination } from 'features/aave/constants'
+import { formatAmount, formatDecimalAsPercent } from 'helpers/formatters/format'
 
 export const getStopLossFormatters = ({
   isShort,
@@ -13,8 +14,14 @@ export const getStopLossFormatters = ({
   priceFormat: string
 }) => {
   const denominationToken = isShort ? collateralTokenSymbol : quoteTokenSymbol
-  const leftFormatter = (x: BigNumber) => (x.isZero() ? '-' : formatPercent(x, { precision: 2 }))
+  const leftFormatter = (x: BigNumber) =>
+    x.isZero() ? '-' : formatDecimalAsPercent(x.times(lambdaPercentageDenomination).div(100))
   const rightFormatter = (x: BigNumber) =>
     x.isZero() ? '-' : `${formatAmount(x, denominationToken)} ${priceFormat}`
   return { leftFormatter, rightFormatter }
 }
+
+export const lambdaParsePercentageValueDown = (val: BigNumber) =>
+  val.div(100).div(lambdaPercentageDenomination)
+export const lambdaParsePercentageValueUp = (val: BigNumber) =>
+  val.times(100).times(lambdaPercentageDenomination)
