+++ features/omni-kit/automation/helpers/index.ts
@@ -3,3 +3,4 @@ export * from './hasActiveProtection'
 export * from './resolveActiveOrder'
 export * from './getRealizedProfit'
 export * from './automationFormatters'
+export * from './useOmniStopLossDataHandler'
